[BROOKLYN-183] Allow use of OSGi framework provided by karaf
Test if we are running inside an OSGi container and use container
services instead of relying on manifest resource on classpath.
Defer to previous implementation when not running inside a container.

Disable failing test on duplicate bundles, since the bundle installation
is subject to change during osgification; it will be handled by karaf.
